backend2/requirements.txt
absl-py==1.0.0
astor==0.8.1
cached-property==1.5.2
certifi==2020.6.20
charset-normalizer==2.0.8
click==8.0.3
dataclasses==0.8
Flask==1.1.2
gast==0.2.2
google-pasta==0.2.0
grpcio==1.42.0
gTTS==2.2.3
gunicorn==20.1.0
h5py==2.10.0
idna==3.3
importlib-metadata==4.8.2
itsdangerous==2.0.1
Jinja2==3.0.3
jsonpickle==2.0.0
Keras==2.2.5
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
Markdown==3.3.6
MarkupSafe==2.0.1
numpy==1.19.5
opt-einsum==3.3.0
Pillow==8.4.0
protobuf==3.19.1
PyYAML==6.0
requests==2.26.0
scipy==1.5.4
six==1.16.0
tensorboard==1.15.0
tensorflow==1.15.0
tensorflow-estimator==1.15.1
termcolor==1.1.0
typing_extensions==4.0.1
urllib3==1.26.7
Werkzeug==2.0.2
wrapt==1.13.3
zipp==3.6.0
